+++ remote/servers.go
@@ -6,6 +6,8 @@ import (
 	"strconv"
 	"sync"
 
+	"github.com/pterodactyl/wings/internal/models"
+
 	"emperror.dev/errors"
 	"github.com/apex/log"
 	"golang.org/x/sync/errgroup"
@@ -173,6 +175,16 @@ func (c *client) SendRestorationStatus(ctx context.Context, backup string, succe
 	resp, err := c.Post(ctx, fmt.Sprintf("/backups/%s/restore", backup), d{"successful": successful})
 	if err != nil {
 		return err
+	}
+	_ = resp.Body.Close()
+	return nil
+}
+
+// SendActivityLogs sends activity logs back to the Panel for processing.
+func (c *client) SendActivityLogs(ctx context.Context, activity []models.Activity) error {
+	resp, err := c.Post(ctx, "/activity", d{"data": activity})
+	if err != nil {
+		return errors.WithStackIf(err)
 	}
 	_ = resp.Body.Close()
 	return nil